Major Changes
Stopped trying to do the timer in the class - I could get it to update
within the class, but couldn’t get it to update the label in the active
view controller.
Started adding all the different view controllers - one for each
screen. This feels weird. I’m repeating so much code that it’s getting
hard to read.
Added alerts for the segues - I wasn’t able to do a yes/no for this.
For some reason, the yes would never work. It was throwing this weird
‘view screen resize doesn’t fit’ or some such. As long as the dialog is
only ‘no’ this works.
There has GOT to be a way to do the timer in the class. I’m going to
post this on some boards and see what I can find out.
At this point, I’m resigned to finish the app as-is, then change it
later.