Oh my...
Well, it’s sort of working.
I took the code out of the view controller updateStopwatch and I used
that function to call newTimer.updateStopwatch() - with newTimer as a
StopWatch2().
This didn’t work as the label was blanked.
So then, I populated the label with 8:00 in viewdidload and viola!
Not really sure how it’s working, but it is :)