Lots of changes
I’ve been working on doing the transition I want. Prior to code I found
online, the transition was only one of the standard five.
I’ve found a way to make it go from side to side.

After that, I had trouble getting the alerts to work. The system was
sending me to the next/previous screen while the timer was still
running. The alert box would pop up, but the app would move anyway.

I had to add a ‘return false’ to my alert. Should have add that from
the beginning.

So, I think this will do it - I just have to update all the view
controllers to the latest code. Famous last words…