Amazing changes...
OK, maybe that’s hyperbole, but it’s still cool.
I’ve got the new transitions working everywhere on the LD side.
When going forward, there’s a ‘show’ segue.
When going back, it uses an exit. Whether or not it needs to do this, I
don’t know.
—goes and tests—
Well, I don’t think it needs to be this way, but I’m going to leave it
because the transition is working. This means that it scrolls left when
clicking the back button. I’m sure there’s a way I can fix this in the
code, but it’s time to get this thing into the app store.
I also added a label at the top and got rid of the nav bar.